MapRoot instead of component/Map
